updated distance calculation logic

+++ src/main/java/com/palriwal/yash/service/flagPoint/ComputeFlagCoordinateService.java
@@ -3,5 +3,5 @@ package com.palriwal.yash.service.flagPoint;
 import com.palriwal.yash.dto.LatLng;
 
 public interface ComputeFlagCoordinateService {
-    LatLng getFlagLatLng(LatLng startPoint, LatLng endPoint, Long distanceFromStart, Long stepDistance);
+    LatLng getFlagLatLng(LatLng startPoint, LatLng endPoint, Double distanceFromStart, Double stepDistance);
 }
