+++ src/main/java/com/palriwal/yash/service/flagPoint/FlagPointService.java
@@ -5,5 +5,5 @@ import com.palriwal.yash.dto.LatLng;
 import java.util.List;
 
 public interface FlagPointService {
-    List<LatLng> getFlagLatLngOnPath(List<com.google.maps.model.LatLng> wayPoints, LatLng startPoint, LatLng endPoint, Long flagDistance);
+    List<LatLng> getFlagLatLngOnPath(List<com.google.maps.model.LatLng> wayPoints, LatLng startPoint, LatLng endPoint, Double flagDistance);
 }
